Added parameter seed to the widget.
I have added a 'seed' parameter to the wordcloud2 function. If it is NULL,
the function still works as expected. In case it does not, the WordCloud
JS function will use a special random generator initialized from a given
seed. This can be useful whenever you want to recreate the same word cloud.

I have borrowed the code for the random number generator from:

https://stackoverflow.com/a/19303725/2061043